pay logic -> only API essentials (bug 1130477)
