--- webpay/bango/templates/bango/cancel.html
@@ -1,12 +1,0 @@
-{% extends "base.html" %}
-
-{% block body_attrs -%}
-  data-cancelflow=true
-{% endblock %}
-
-{% block content %}
-  <div id="progress" class="progress">
-    <div class="throbber"></div>
-    <span class="txt">{{ _('Payment cancelled.') }}</span>
-  </div>
-{% endblock %}
