--- webpay/base/templates/base.html
@@ -1,84 +1,0 @@
-<!DOCTYPE html>
-<html LANG="{{ LANG }}" dir="{{ DIR }}">
-  <head>
-    <meta content="text/html; charset=UTF-8" http-equiv="Content-Type">
-    <meta name="csrf" content="{{ csrf_token }}">
-    <meta name="viewport" content="width=device-width, user-scalable=no, initial-scale=1, maximum-scale=1">
-    <meta name="apple-mobile-web-app-capable" content="yes">
-    {# L10n: Web Pay is what we call the payment server/flow. #}
-    <title>{% block page_title %}{{ _('Web Pay') }}{% endblock %}</title>
-
-    {% block extra_head %}
-    {% endblock %}
-
-  </head>
-  <body
-    {% block body_attrs %}{% endblock %}
-    data-privacy-policy="https://marketplace.firefox.com/privacy-policy"
-    data-terms-of-service="https://marketplace.firefox.com/terms-of-use"
-    data-unverified-issuer="{{ settings.BROWSERID_UNVERIFIED_ISSUER }}"
-    data-logged-in-user="{{ session.get('logged_in_user', '') }}"
-    data-reset-user-url="{{ url('auth.reset_user') }}"
-    data-denied-url="{{ url('auth.denied') }}"
-    data-bango-logout-url="{{ settings.PAY_URLS[settings.PAYMENT_PROVIDER]['base'] + settings.PAY_URLS[settings.PAYMENT_PROVIDER]['logout'] }}"
-    data-pin-num-chars-warning="{{ _('Pin must be 4 digits.') }}"
-    data-pin-non-numeric-warning="{{ _('Pin can only contain digits.') }}"
-    data-pin-reauth-msg="{{ _('You logged in as a different user, you must start the payment again.') }}"
-    data-loading-msg="{{ _('Loading') }}"
-    data-begin-msg="{{ _('Beginning payment&hellip;') }}"
-    data-persona-msg="{{ _('Connecting to Persona') }}"
-    data-complete-msg="{{ _('Payment complete') }}"
-    data-cancelled-msg="{{ _('Payment cancelled.') }}"
-    data-logout-timeout="{{ settings.LOGOUT_TIMEOUT }}"
-    data-login-timeout="{{ settings.LOGIN_TIMEOUT }}"
-    data-settings="{{ settings.JS_SETTINGS|json }}"
-    data-error-code="{{ error_code }}"
-    data-cancel-code="{{ dev_messages.USER_CANCELLED }}"
-    data-full-error-heading="{{ _('Oops&hellip;') }}"
-    data-full-error-detail="{{ _('This is taking longer than expected. Try again?') }}"
-    data-full-error-detail-generic="{{ _('Something went wrong. Try again?') }}"
-    data-full-error-confirm="{{ _('OK') }}"
-    data-full-error-cancel="{{ _('Cancel') }}"
-    data-static-url="{{ settings.STATIC_URL }}"
-    data-configure-transaction="{{ url('pay.configure_transaction') }}"
-    data-trans-failed-msg="{{ _('The transaction failed. You have not been charged for this purchase.') }}"
-    data-trans-status-completed="{{ solitude_constants.STATUS_COMPLETED }}"
-    data-trans-status-pending="{{ solitude_constants.STATUS_PENDING }}"
-    data-trans-status-failed="{{ solitude_constants.STATUS_FAILED }}"
-    data-trans-status-cancelled="{{ solitude_constants.STATUS_CANCELLED }}"
-    >
-    <section class="pay">
-      <div id="content">
-        {% block extra_content %}{% endblock %}
-        {% block content %}{% endblock %}
-        {% if messages %}
-        <ul class="messages">
-           {% for message in messages %}
-           <li><p>{{ message }}</p></li>
-           {% endfor %}
-        </ul>
-        {% endif %}
-
-        {# Full screen error html #}
-        <div id="full-screen-error" class="hidden">
-          <div class="msg">
-            <h2 class="heading"></h2>
-            <p class="detail"></p>
-            <p class="error-code"></p>
-          </div>
-          <footer>
-            <a href="#" class="button cancel sec"></a>
-            <a href="#" class="button confirm"></a>
-          </footer>
-        </div>
-
-      </div>
-    </section>
-
-    {# Progress indicator html #}
-    <div id="progress" class="progress hidden hidden-full-error">
-      <div class="throbber"></div>
-      <span class="txt"></span>
-    </div>
-  </body>
-</html>
