--- webpay/base/templates/error.html
@@ -1,20 +1,0 @@
-{% extends "base.html" %}
-
-{% block body_attrs -%}
-  class="full-error"
-{%- endblock %}
-
-{% block content %}
-  <div class="msg">
-    <h2>{{ _('Error') }}</h2>
-    <p>{{ error }}</p>
-    {% if error_code %}
-      <p class="error-code">{{ error_code }}</p>
-    {% endif %}
-  </div>
-  <footer>
-    <a class="single button cancel" href="#">
-      {{ _('Cancel') }}
-    </a>
-  </footer>
-{% endblock %}
