--- webpay/pay/templates/pay/fake-bango-url.html
@@ -1,12 +1,0 @@
-{% extends "base.html" %}
-
-{% block content %}
-  <div class="message">
-    <p>
-      Pretend this is a Bango payment screen for configuration ID {{ bill_config_id }}
-    </p>
-    <p>
-      <a href="{{ url('bango.success') }}">Complete the payment</a>
-    </p>
-  </div>
-{% endblock %}
