--- webpay/pay/templates/pay/lobby.html
@@ -1,15 +1,0 @@
-{% extends "pin/pin_form.html" %}
-
-{% block body_attrs -%}
-  data-flow="lobby"
-  data-verify-url="{{ url('auth.verify') }}"
-  data-pin-error-codes="{{ PIN_ERROR_CODES|json }}"
-{%- endblock %}
-
-{% block extra_content %}
-  <div id="login" class="message hidden hidden-full-error">
-    <h2>{{ _('Sign In') }}</h2>
-    <p>{{ _('Sign in to continue with the payment') }}</p>
-    <a id="signin" class="sign-in" href="#">{{ _('Sign in') }}</a>
-  </div>
-{% endblock %}
