--- webpay/pay/templates/pay/simulate_done.html
@@ -1,10 +1,0 @@
-{% extends "base.html" %}
-
-{% block body_attrs -%}
-  data-docomplete="true"
-{%- endblock %}
-
-{% block content %}
-  <h2>{{ _('Simulate Payment') }}</h2>
-  <p class="centered">{{ _('Your payment is complete.') }}</p>
-{% endblock %}
