--- webpay/pay/templates/pay/super_simulate.html
@@ -1,20 +1,0 @@
-{% extends "base.html" %}
-
-{% block content %}
-  <h2 class="centered">{{ _('You have super powers') }}</h2>
-  <form id="super-simulate" action="{{ url('pay.super_simulate') }}" method="post">
-    {{ csrf() }}
-    {{ form.non_field_errors() }}
-    <p>
-      {{ form.action.errors }}
-      {{ form.action }}
-    </p>
-    <p>
-      {{ form.network.errors }}
-      {{ form.network }}
-    </p>
-    <p>
-      <input class="button" type="submit" value="{{ _('Continue') }}">
-    </p>
-  </form>
-{% endblock %}
