--- webpay/pay/templates/pay/wait-to-start.html
@@ -1,14 +1,0 @@
-{% extends "base.html" %}
-
-{% block body_attrs -%}
-  data-flow="wait"
-  data-trans-start-url="{{ url('pay.trans_start_url') }}"
-  data-loading-msg="{{ _('Setting up payment&hellip;') }}"
-{%- endblock %}
-
-{% block content %}
-<div id="progress" class="progress">
-  <div class="throbber"></div>
-  <span class="txt">{{ _('Setting up payment&hellip;') }}</span>
-</div>
-{% endblock %}
