+++ webpay/pay/views.py
@@ -25,7 +25,7 @@ from webpay.pin.forms import VerifyPinForm
 
 from lib.marketplace.api import client as marketplace, UnknownPricePoint
 from lib.solitude import constants
-from lib.solitude.api import client as solitude, ProviderHelper
+from lib.solitude.api import client as solitude
 from lib.solitude.exceptions import ResourceModified
 
 from . import tasks
@@ -181,321 +181,6 @@ def configure_transaction(request, data=None):
             'payment_required': payment_required}
 
 
-def index(request):
-    """Hand off either lobby or serving Spartacus depending on settings."""
-    if settings.SPA_ENABLE:
-        from webpay.spa.views import index as spa_index
-        return spa_index(request)
-    else:
-        return lobby(request)
-
-
-@anonymous_csrf_exempt
-@require_GET
-def lobby(request):
-    sess = request.session
-    have_jwt = bool(request.GET.get('req'))
-
-    log.info('starting from JWT? {have_jwt}'.format(have_jwt=have_jwt))
-    if have_jwt:
-        # If it returns a response there was likely
-        # an error and we should return it.
-        res = process_pay_req(request)
-        if isinstance(res, http.HttpResponse):
-            return res
-    elif settings.TEST_PIN_UI:
-        # This won't get you very far but it lets you create/enter PINs
-        # and stops a traceback after that.
-        sess['trans_id'] = trans_id()
-
-    pin_form = VerifyPinForm()
-
-    if sess.get('uuid'):
-        auth_utils.update_session(request,
-                                  sess.get('uuid'),
-                                  False,
-                                  request.session.get('logged_in_user', None))
-
-        redirect_url = check_pin_status(request)
-        if redirect_url is not None:
-            return http.HttpResponseRedirect(
-                '{0}?next={1}'.format(reverse('pay.bounce'), redirect_url)
-            )
-
-    # If the buyer closed the trusted UI during reset flow, we want to unset
-    # the reset pin flag. They can hit the forgot pin button if they still
-    # don't remember their pin.
-    if sess.get('uuid_needs_pin_reset'):
-        try:
-            solitude.set_needs_pin_reset(sess['uuid'], False)
-        except ResourceModified:
-            return system_error(request, code=msg.RESOURCE_MODIFIED)
-        sess['uuid_needs_pin_reset'] = False
-
-    if sess.get('is_simulation', False):
-        sim_req = sess['notes']['pay_request']['request']['simulate']
-        log.info('Starting simulate %s for %s'
-                 % (sim_req, sess['notes']['issuer_key']))
-        return render(request, 'pay/simulate.html', {
-            'simulate': sim_req
-        })
-
-    return render(request, 'pay/lobby.html', {
-        'action': reverse('pin.verify'),
-        'form': pin_form,
-        'title': _('Enter Pin'),
-        'track_cancel': {
-            'action': 'pin cancel',
-            'label': 'Pin Entry Page',
-        },
-    })
-
-
-@require_GET
-def bounce(request):
-    next = request.GET.get('next')
-    if not next or not next.startswith('/mozpay/'):
-        return http.HttpResponseForbidden()
-    return render(request, 'pay/bounce.html', {
-        'next': next
-    })
-
-
-@require_POST
-def simulate(request):
-    if not request.session.get('is_simulation', False):
-        log.info('Request to simulate without a valid session')
-        return http.HttpResponseForbidden()
-
-    tasks.simulate_notify.delay(request.session['notes']['issuer_key'],
-                                request.session['notes']['pay_request'])
-    return render(request, 'pay/simulate_done.html', {})
-
-
-@user_can_simulate
-def super_simulate(request):
-    if not settings.ALLOW_ADMIN_SIMULATIONS:
-        return http.HttpResponseForbidden()
-
-    form = SuperSimulateForm(request.POST or None)
-    if request.method == 'POST' and form.is_valid():
-
-        if form.cleaned_data['action'] == 'real':
-            if form.cleaned_data['network']:
-                mcc, mnc = form.cleaned_data['network']
-                was_configured, error_code = reconfigure_transaction(
-                    request, mcc, mnc)
-                if not was_configured:
-                    if not error_code:
-                        error_code = msg.TRANS_CONFIG_FAILED
-                    log.error('Re-configuring transaction failed: {er}'
-                              .format(er=error_code))
-                    return system_error(request, code=error_code)
-
-            # Continue to the wait screen.
-            return redirect(reverse('pay.wait_to_start'))
-
-        elif form.cleaned_data['action'] == 'simulate':
-            send_simulated_notification(request.session['trans_id'])
-            return render(request, 'pay/simulate_done.html', {})
-
-        else:
-            raise ValueError('Not sure what to do for action {a}'
-                             .format(a=form.cleaned_data['action']))
-
-    return render(request, 'pay/super_simulate.html', {'form': form})
-
-
-def reconfigure_transaction(request, mcc, mnc):
-    """Configure a new transaction with a simulated network."""
-    new_trans_id = trans_id()
-    log.warning('Replacing transaction for simulation. '
-                'Old: {old}; new: {new}'
-                .format(old=request.session['trans_id'],
-                        new=new_trans_id))
-    # TODO: patch the old transaction to mark it cancelled?
-    request.session['trans_id'] = new_trans_id
-
-    log.info('Simulating a network for transaction {t}: '
-             'mcc={mcc}; mnc={mnc}'
-             .format(mcc=mcc, mnc=mnc, t=request.session['trans_id']))
-    return tasks.configure_transaction(request, mcc=mcc, mnc=mnc)
-
-
-def send_simulated_notification(trans_id):
-    try:
-        trans = solitude.get_transaction(trans_id)
-    except ObjectDoesNotExist:
-        # This could mean:
-        # * the transaction was never configured due to error.
-        # * the celery configuration task was just too slow.
-        raise ValueError(
-            'Cannot simulate transaction {t}, not configured'
-            .format(t=trans_id))
-    # TODO: patch solitude to mark this as a super-simulated
-    # transaction.
-
-    req = trans['notes']['pay_request']
-    # TODO: support simulating refunds.
-    req['request']['simulate'] = {'result': 'postback'}
-    tasks.simulate_notify.delay(trans['notes']['issuer_key'], req)
-
-
-@user_verified
-@require_GET
-def fake_bango_url(request):
-    return render(request, 'pay/fake-bango-url.html',
-                  {'bill_config_id': request.GET['bcid']})
-
-
-@user_verified
-@require_GET
-def wait_to_start(request):
-    """
-    Wait until the transaction is in a ready state.
-
-    The transaction was started previously during the buy flow in the
-    background from webpay.pay.tasks.
-
-    Serve JS that polls for transaction state.
-    When ready, redirect to the Bango payment URL using
-    the generated billing configuration ID.
-    """
-    trans_id = request.session.get('trans_id', None)
-    if not trans_id:
-        # This seems like a seriously problem but maybe there is just a race
-        # condition. If we see a lot of these in the logs it means the
-        # payment will never complete so we should keep an eye on it.
-        log.error('wait_to_start() session trans_id {t} was None'
-                  .format(t=trans_id))
-    try:
-        statsd.incr('purchase.payment_time.retry')
-        with statsd.timer('purchase.payment_time.get_transation'):
-            trans = solitude.get_transaction(trans_id)
-    except ObjectDoesNotExist:
-        trans = {'status': None}
-
-    if trans['status'] in constants.STATUS_ENDED:
-        statsd.incr('purchase.payment_time.failure')
-        log.exception('Attempt to restart finished transaction {0} '
-                      'with status {1}'.format(trans_id, trans['status']))
-        return system_error(request, code=msg.TRANS_ENDED)
-
-    if trans['status'] == constants.STATUS_PENDING:
-        statsd.incr('purchase.payment_time.success')
-        payment_start = request.session.get('payment_start', False)
-        if payment_start:
-            delta = int((time.time() - float(payment_start)) * 1000)
-            statsd.timing('purchase.payment_time.duration', delta)
-        # Dump any messages so we don't show them later.
-        clear_messages(request)
-        # The transaction is ready; no need to wait for it.
-        url = get_payment_url(trans)
-        log.info('immediately redirecting to payment URL {url} '
-                 'for trans {tr}'.format(url=url, tr=trans))
-        return http.HttpResponseRedirect(url)
-
-    return render(request, 'pay/wait-to-start.html')
-
-
-@user_verified
-@json_view
-@require_GET
-def trans_start_url(request):
-    """
-    JSON handler to get the Bango payment URL to start a transaction.
-    """
-    trans = None
-    trans_id = request.session.get('trans_id')
-    data = {'url': None, 'status': None, 'provider': None}
-
-    if not trans_id:
-        log.error('trans_start_url(): no transaction ID in session')
-        return http.HttpResponseBadRequest()
-    try:
-        statsd.incr('purchase.payment_time.retry')
-        with statsd.timer('purchase.payment_time.get_transaction'):
-            trans = solitude.get_transaction(trans_id)
-        data['status'] = trans['status']
-        data['provider'] = constants.PROVIDERS_INVERTED.get(trans['provider'])
-    except ObjectDoesNotExist:
-        log.error('trans_start_url() transaction does not exist: {t}'
-                  .format(t=trans_id))
-
-    if data['status'] == constants.STATUS_PENDING:
-        statsd.incr('purchase.payment_time.success')
-        payment_start = request.session.get('payment_start', False)
-        if payment_start:
-            delta = int((time.time() - float(payment_start)) * 1000)
-            statsd.timing('purchase.payment_time.duration', delta)
-        url = get_payment_url(trans)
-        log.info('async call got payment URL {url} for trans {tr}'
-                 .format(url=url, tr=trans))
-        data['url'] = url
-
-    if trans and trans['status'] == constants.STATUS_ERRORED:
-        statsd.incr('purchase.payment_time.errored')
-        log.exception('Purchase configuration failed: {0} with status {1}'
-                      .format(trans_id, trans['status']))
-        return system_error(
-            request,
-            code=getattr(msg, trans.get('status_reason', 'UNEXPECTED_ERROR'))
-        )
-
-    return data
-
-
-def _callback_url(request, is_success):
-    status = is_success and 'success' or 'error'
-    signed_notice = request.POST['signed_notice']
-    statsd.incr('purchase.payment_{0}_callback.received'.format(status))
-
-    # This is currently only used by Bango and Zippy.
-    # Future providers should probably get added to the notification
-    # abstraction in provider/views.py
-    provider = ProviderHelper(settings.PAYMENT_PROVIDER)
-
-    if provider.is_callback_token_valid(signed_notice):
-        statsd.incr('purchase.payment_{0}_callback.ok'.format(status))
-        log.info('Callback {0} token was valid.'.format(status))
-        querystring = http.QueryDict(signed_notice)
-        if 'ext_transaction_id' in querystring:
-            ext_transaction_id = querystring['ext_transaction_id']
-            if is_success:
-                tasks.payment_notify.delay(ext_transaction_id)
-            else:
-                tasks.chargeback_notify.delay(ext_transaction_id)
-            return http.HttpResponse(status=204)
-        else:
-            statsd.incr('purchase.payment_{0}_callback.incomplete'
-                        ''.format(status))
-            log.error('Callback {0} token was incomplete: '
-                      '{1}'.format(status, querystring))
-    else:
-        statsd.incr('purchase.payment_{0}_callback.fail'.format(status))
-        log.error('Callback {0} token was invalid: '
-                  '{1}'.format(status, signed_notice))
-    return http.HttpResponseBadRequest()
-
-
-@require_POST
-def callback_success_url(request):
-    """
-    Hit by the provider to confirm a success of a transaction
-    in case of an interruption in the payment process.
-    """
-    return _callback_url(request, is_success=True)
-
-
-@require_POST
-def callback_error_url(request):
-    """
-    Hit by the provider to confirm a failure of a transaction
-    in case of an interruption in the payment process.
-    """
-    return _callback_url(request, is_success=False)
-
-
 def _trim_pay_request(req):
 
     def _trim(st):
@@ -541,18 +226,3 @@ def disabled_by_user_agent(user_agent):
                  .format(ua=user_agent))
 
     return is_disabled
-
-
-def get_payment_url(transaction):
-    """
-    Given a Solitude transaction object, return a URL to start the payment
-    flow for this provider.
-    """
-    # A transaction pay_url is configured at the time that a
-    # transaction is started.
-    url = transaction['pay_url']
-    log.info('Start pay provider payflow "{pr}" for '
-             'transaction {tr} at: {url}'
-             .format(pr=transaction.get('provider'), url=url,
-                     tr=transaction.get('uuid')))
-    return url
