--- webpay/provider/templates/provider/success.html
@@ -1,12 +1,0 @@
-{% extends "base.html" %}
-
-{% block body_attrs -%}
-  data-docomplete="true"
-{%- endblock %}
-
-{% block content %}
-  <div id="progress" class="progress">
-    <div class="throbber"></div>
-    <span class="txt">{{ _('Your payment is complete.') }}</span>
-  </div>
-{% endblock %}
