--- webpay/provider/templates/provider/wait-to-finish.html
@@ -1,14 +1,0 @@
-{% extends "base.html" %}
-
-{% block body_attrs -%}
-  data-flow="wait-to-finish"
-  data-trans-start-url="{{ transaction_status_url }}"
-  data-loading-msg="{{ _('Verifying payment&hellip;') }}"
-{%- endblock %}
-
-{% block content %}
-<div id="progress" class="progress">
-  <div class="throbber"></div>
-  <span class="txt">{{ _('Verifying payment&hellip;') }}</span>
-</div>
-{% endblock %}
